requirements.txt
Flask
Flask-WTF
WTForms
gunicorn
python-dotenv
Flask-Admin
Flask-SQLAlchemy
Flask-Login